Refactor services into CoreModule
