+++ src/app/core/contents-xml.service.spec.ts
