+++ src/app/core/contents-xml.service.ts
