+++ src/app/core/contributor.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed, getTestBed, inject } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 
 import { ContributorService } from './contributor.service';
-import { environment } from '../environments/environment';
+import { environment } from '../../environments/environment';
 
 describe('ContributorService', () => {
   let httpClient: HttpTestingController,
