+++ src/app/event-page/event-depth.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { FormatterService } from '../formatter.service';
+import { FormatterService } from '../core/formatter.service';
 
 @Pipe({
   name: 'eventDepth'
