+++ src/app/event-page/footer/footer.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 
-import { ContributorService } from '../../contributor.service';
+import { ContributorService } from '../../core/contributor.service';
 
 
 @Component({
