+++ src/app/moment-tensor/info/info.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { InfoComponent } from './info.component';
 import { MockComponent } from 'ng2-mock-component';
-import { FormatterService } from '../../formatter.service';
+import { FormatterService } from '../../core/formatter.service';
 import { MockPipe } from '../../mock-pipe';
 
 describe('InfoComponent', () => {
