+++ src/app/product-page/date-time.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { FormatterService } from '../formatter.service';
+import { FormatterService } from '../core/formatter.service';
 
 @Pipe({
   name: 'dateTime'
