+++ src/app/shared/coordinates/coordinates.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CoordinatesComponent } from './coordinates.component';
 
-import { FormatterService } from '../../formatter.service';
+import { FormatterService } from '../../core/formatter.service';
 
 
 describe('CoordinatesComponent', () => {
