+++ src/app/shared/fe-region/fe-region.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { FeRegionComponent } from './fe-region.component';
 
-import { GeoserveService } from '../../geoserve.service';
+import { GeoserveService } from '../../core/geoserve.service';
 
 
 describe('FeRegionComponent', () => {
