+++ src/app/technical/moment-tensor-summary/moment-tensor-summary.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 
 import { Tensor } from '../../shared/beachball/tensor';
-import { FormatterService } from '../../formatter.service';
+import { FormatterService } from '../../core/formatter.service';
 
 @Component({
   selector: 'technical-moment-tensor-summary',
