+++ src/app/technical/technical/technical.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { EventService } from '../../event.service';
+import { EventService } from '../../core/event.service';
 
 @Component({
   selector: 'technical',
